requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
drf-yasg
psycopg2-binary